--- src/main/java/SYSC6/Project/sorting/Direction.java
@@ -1,6 +1,0 @@
-package SYSC6.Project.sorting;
-
-public enum Direction {
-    asc,
-    desc;
-}
