+++ src/main/java/SYSC6/Project/user/userManagement/UserService.java
@@ -1,29 +1,32 @@
 package SYSC6.Project.user.userManagement;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import SYSC6.Project.sorting.*;
 import SYSC6.Project.user.UserRepository;
+import SYSC6.Project.user.userManagement.sorting.Column;
+import SYSC6.Project.user.userManagement.sorting.Order;
+import SYSC6.Project.user.userManagement.sorting.Page;
+import SYSC6.Project.user.userManagement.sorting.PagingRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import lombok.extern.slf4j.Slf4j;
 
+/**
+ * Used for generating Page objects of UserPOJO's for use populating dynamic tables
+ */
 @Slf4j
 @Service
 public class UserService {
 
     private static final Comparator<UserPOJO> EMPTY_COMPARATOR = (e1, e2) -> 0;
 
-    private static final SimpleDateFormat sdf = new SimpleDateFormat("yyyy/MM/dd");
-
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
 
     @Autowired
     public UserService(UserRepository userRepository) {
@@ -32,7 +35,7 @@ public class UserService {
 
     public Page<UserPOJO> getUsers(PagingRequest pagingRequest) {
 
-        List<UserPOJO> users = new ArrayList<UserPOJO>();
+        List<UserPOJO> users = new ArrayList<>();
         userRepository.findAll().forEach(u -> users.add(u.toPojo()));
         return getPage(users, pagingRequest);
 
@@ -70,7 +73,7 @@ public class UserService {
         return user -> user.getUsername()
                 .toLowerCase()
                 .contains(value)
-                || user.getRole().toString()
+                || user.getRole()
                 .toLowerCase()
                 .contains(value);
     }
