+++ src/main/java/SYSC6/Project/user/userManagement/sorting/Direction.java
@@ -1,0 +1,6 @@
+package SYSC6.Project.user.userManagement.sorting;
+
+public enum Direction {
+    asc,
+    desc;
+}
