+++ src/main/resources/templates/Registration.html
@@ -19,7 +19,7 @@
             <div class="col-md-6 col-lg-4">
                 <div class="login-wrap p-0">
                     <h3 class="mb-4 text-center">Password must contain:</h3>
-                    <h3 class="mb-4 text-center">At least 6 Characters</h3>
+                    <h3 class="mb-4 text-center">At least 7 Characters</h3>
                     <h3 class="mb-4 text-center">At least 1 Number</h3>
                     <h3 class="mb-4 text-center">At least 1 Special Character</h3>
                     <form id="hello" action="#" th:action="@{/TempCreate}" method="post">
